requirements.txt
langchain
langchain-community
langchain-huggingface
sentence-transformers
pypdf
python-dotenv
faiss-cpu
openai
streamlit